rest api update, some dao updates

+++ src/main/java/com/calendar/project/controller/JSONController.java
@@ -4,6 +4,7 @@ import com.calendar.project.exception.FirebaseException;
 import com.calendar.project.exception.JacksonUtilityException;
 import com.calendar.project.model.*;
 import com.calendar.project.model.enums.EventType;
+import com.calendar.project.model.enums.TagType;
 import com.calendar.project.service.*;
 import com.calendar.project.model.dto.EventResource;
 import com.calendar.project.service.impl.Firebase;
@@ -12,15 +13,26 @@ import com.calendar.project.service.UserService;
 import com.calendar.project.service.EventService;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
+import com.sun.org.apache.xml.internal.security.utils.Base64;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import com.calendar.project.model.User;
+import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.util.*;
@@ -46,6 +58,9 @@ public class JSONController {
 
     @Autowired
     NotificationService notificationService;
+
+    @Autowired
+    private AuthenticationManager authenticationManager;
 
 
 //    @RequestMapping(value = "/sendToFirebase",method = RequestMethod.GET)
@@ -119,7 +134,6 @@ public class JSONController {
         event.setAuthor(securityService.findLoggedInUsername());
         eventService.saveEvent(event);
     }
-
     @RequestMapping(value = "/updateEventJson", params = "id", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json")
     @ResponseStatus(HttpStatus.OK)
     public @ResponseBody ResponseEntity updateEvent(@PathVariable @RequestParam("id") int id, @RequestBody EventResource eventResource) {
@@ -135,10 +149,14 @@ public class JSONController {
 
     @RequestMapping(value = "/deleteEventJson", params = {"id"}, method = RequestMethod.DELETE, consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json")
     @ResponseStatus(HttpStatus.OK)
-    public @ResponseBody void deleteEvent(@PathVariable @RequestParam("id") int id) {
+    public @ResponseBody ResponseEntity deleteEvent(@PathVariable @RequestParam("id") int id) {
+        User user = securityService.findLoggedInUsername();
         Event event = eventService.getEvent(id);
-//        Event event = Converter.convert(eventResource);
+        if ((!user.getId().equals(event.getAuthor().getId()))&&
+                (!user.getId().equals(userService.findByUsername("admin").getId())))
+            return new ResponseEntity(HttpStatus.METHOD_NOT_ALLOWED);
         eventService.deleteEvent(event);
+        return new ResponseEntity(HttpStatus.OK);
     }
 
     @GetMapping(value = "/allTags", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -187,17 +205,85 @@ public class JSONController {
 
     @RequestMapping(value = "/editUserJson", params = "username", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json")
     @ResponseStatus(HttpStatus.OK)
-    public @ResponseBody void editUser(@PathVariable @RequestParam("username") String username, @RequestBody User user) {
+    public @ResponseBody ResponseEntity editUser(@PathVariable @RequestParam("username") String username, @RequestBody User user) {
         User firstUser = userService.findByUsername(username);
         User userFinal = userService.updateUserForRest(firstUser, user);
+        User user1 = securityService.findLoggedInUsername();
+        if ((!user1.getId().equals(userFinal.getId()))&&
+                (!user1.getId().equals(userService.findByUsername("admin").getId())))
+            return new ResponseEntity(HttpStatus.METHOD_NOT_ALLOWED);
         userService.update(userFinal);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
-//    @RequestMapping(value = "/loginJson", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json")
-//    @ResponseStatus(HttpStatus.CREATED)
-//    public @ResponseBody void loginJson(@RequestBody User user) {
-//
-//    }
+    @RequestMapping(value = "/createUserJson", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json")
+    @ResponseStatus(HttpStatus.OK)
+    public @ResponseBody void createUser(@RequestBody User user) throws IOException {
+        MultipartFile userImage = user.getMultipartFile();
+        if (userImage == null) userService.save(user);
+        else{
+        try{
+            user.setImage(Base64.encode(userImage.getBytes()));
+        } catch(IOException e){
+            e.printStackTrace();
+        }
+        userService.save(user);}
+    }
 
+    @GetMapping(value = "/getEventsByUser", params = {"id"}, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> getEventsByUser(@PathVariable @RequestParam("id") long id) throws IOException {
+        List<Event> events = eventService.getEventsByUser(id);
+        String eventString = eventService.getEventsJson(events);
+        return new ResponseEntity<>(eventString, HttpStatus.OK);
+    }
+
+    @GetMapping(value = "/getEventsByAuthor", params = {"id"}, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> getEventsByAuthor(@PathVariable @RequestParam("id") long id) throws IOException {
+        List<Event> events = eventService.getEventsByAuthor(id);
+        String eventString = eventService.getEventsJson(events);
+        return new ResponseEntity<>(eventString, HttpStatus.OK);
+    }
+
+    @GetMapping(value = "/getEventsByTag", params = {"tag"}, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> getEventsByTag(@PathVariable @RequestParam("tag") TagType tag) throws IOException {
+        List<Event> events = eventService.getEventsByTag(tag);
+        String eventString = eventService.getEventsJson(events);
+        return new ResponseEntity<>(eventString, HttpStatus.OK);
+    }
+
+    @GetMapping(value = "/getEventsByType", params = {"type"}, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> getEventsByType(@PathVariable @RequestParam("type") EventType type) throws IOException {
+        List<Event> events = eventService.getEventsByType(type);
+        String eventString = eventService.getEventsJson(events);
+        return new ResponseEntity<>(eventString, HttpStatus.OK);
+    }
+
+
+
+
+//    @RequestMapping(value = "/login", method =RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = "application/json")
+//    public @ResponseBody String authentication(@PathVariable @RequestParam("login") String username,
+//                                               @PathVariable @RequestParam("password") String password, HttpServletRequest request) {
+//
+//        Authentication authenticationToken = new UsernamePasswordAuthenticationToken(
+//                username, password);
+//        try {
+//
+//            Authentication authentication = authenticationManager
+//                    .authenticate(authenticationToken);
+//
+//
+//            SecurityContext securityContext = SecurityContextHolder
+//                    .getContext();
+//
+//            securityContext.setAuthentication(authentication);
+//
+//            HttpSession session = request.getSession(true);
+//            session.setAttribute("SPRING_SECURITY_CONTEXT", securityContext);
+//
+//            return "sucess";
+//        } catch (AuthenticationException ex) {
+//            return "fail " + ex.getMessage();
+//        }
 
 }
